pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "fluxqueue"
dynamic = ["version"]
description = "A blazingly fast, lightweight task queue for Python, written in Rust"
readme = "README.md"
requires-python = ">=3.11,<3.15"
keywords = ["task-queue", "queue", "redis", "async", "background-tasks", "worker", "rust", "fluxqueue"]
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [
    {name = "Giorgi Merebashvili", email = "mereba2627@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache License 2.0",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Rust",
    "Operating System :: POSIX :: Linux",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Distributed Computing",
]

[project.urls]
Repository = "https://github.com/CCXLV/fluxqueue"
Documentation = "https://fluxqueue.ccxlv.dev"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest-asyncio",
    "python-dotenv",
    "redis",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocstrings-python",
    "maturin",
]
tests = [
    "pytest-asyncio",
    "python-dotenv",
    "testcontainers",
    "redis"
]
cli = [
    "fluxqueue-cli>=0.1.0b4"
]
docs = [
    "mkdocs-material", 
    "mkdocstrings",
    "mkdocstrings-python",
]
build = ["maturin"]

[tool.maturin]
python-source = "."
module-name = "fluxqueue.fluxqueue_core"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "C",   # flake8-comprehensions
  "B",   # flake8-bugbear
  "UP",  # pyupgrade
  "N",   # pep8-naming
  "SIM", # flake8-simplify
  "RUF", # ruff-specific rules
]
ignore = [
  "E501", # line too long, handled by black
  "B008", # do not perform function calls in argument defaults
  "C901", # complexity
  "N806", # variable in function should be lowercase (allows CamelCase for pydantic models)
]
exclude = [".git", ".ruff_cache", ".venv"]
